+++ lib/vqf/vqf.h
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-// Modified to add timestamps in: updateGyr(const vqf_real_t gyr[3], double gyrTs)
+// Modified to add timestamps in: updateGyr(const vqf_real_t gyr[3], float gyrTs)
 // Removed batch update functions
 
 #ifndef VQF_HPP
@@ -18,12 +18,12 @@
 /**
  * @brief Typedef for the floating-point data type used for most operations.
  *
- * By default, all floating-point calculations are performed using `double`. Set the `VQF_SINGLE_PRECISION` define to
- * change this type to `float`. Note that the Butterworth filter implementation will always use double precision as
+ * By default, all floating-point calculations are performed using `float`. Set the `VQF_SINGLE_PRECISION` define to
+ * change this type to `float`. Note that the Butterworth filter implementation will always use float precision as
  * using floats can cause numeric issues.
  */
 #ifndef VQF_SINGLE_PRECISION
-typedef double vqf_real_t;
+typedef float vqf_real_t;
 #else
 typedef float vqf_real_t;
 #endif
@@ -321,7 +321,7 @@ struct VQFState {
     /**
      * @brief Internal low-pass filter state for #lastAccLp.
      */
-    double accLpState[3*2];
+    float accLpState[3*2];
     /**
      * @brief Last inclination correction angular rate.
      *
@@ -377,11 +377,11 @@ struct VQFState {
      * @brief Internal state of the Butterworth low-pass filter for the rotation matrix coefficients used in motion
      * bias estimation.
      */
-    double motionBiasEstRLpState[9*2];
+    float motionBiasEstRLpState[9*2];
     /**
      * @brief Internal low-pass filter state for the rotated bias estimate used in motion bias estimation.
      */
-    double motionBiasEstBiasLpState[2*2];
+    float motionBiasEstBiasLpState[2*2];
 #endif
     /**
      * @brief Last (squared) deviations from the reference of the last sample used in rest detection.
@@ -408,7 +408,7 @@ struct VQFState {
     /**
      * @brief Internal low-pass filter state for #restLastGyrLp.
      */
-    double restGyrLpState[3*2];
+    float restGyrLpState[3*2];
     /**
      * @brief Last low-pass filtered accelerometer measurement used as the reference for rest detection.
      */
@@ -416,7 +416,7 @@ struct VQFState {
     /**
      * @brief Internal low-pass filter state for #restLastAccLp.
      */
-    double restAccLpState[3*2];
+    float restAccLpState[3*2];
 
     /**
      * @brief Norm of the currently accepted magnetic field reference.
@@ -465,7 +465,7 @@ struct VQFState {
     /**
      * @brief Internal low-pass filter state for the current norm and dip angle.
      */
-    double magNormDipLpState[2*2];
+    float magNormDipLpState[2*2];
 };
 
 /**
@@ -494,13 +494,13 @@ struct VQFCoefficients
      *
      * The array contains \f$\begin{bmatrix}b_0 & b_1 & b_2\end{bmatrix}\f$.
      */
-    double accLpB[3];
+    float accLpB[3];
     /**
      * @brief Denominator coefficients of the acceleration low-pass filter.
      *
      * The array contains \f$\begin{bmatrix}a_1 & a_2\end{bmatrix}\f$ and \f$a_0=1\f$.
      */
-    double accLpA[2];
+    float accLpA[2];
 
     /**
      * @brief Gain of the first-order filter used for heading correction.
@@ -533,19 +533,19 @@ struct VQFCoefficients
     /**
      * @brief Numerator coefficients of the gyroscope measurement low-pass filter for rest detection.
      */
-    double restGyrLpB[3];
+    float restGyrLpB[3];
     /**
      * @brief Denominator coefficients of the gyroscope measurement low-pass filter for rest detection.
      */
-    double restGyrLpA[2];
+    float restGyrLpA[2];
     /**
      * @brief Numerator coefficients of the accelerometer measurement low-pass filter for rest detection.
      */
-    double restAccLpB[3];
+    float restAccLpB[3];
     /**
      * @brief Denominator coefficients of the accelerometer measurement low-pass filter for rest detection.
      */
-    double restAccLpA[2];
+    float restAccLpA[2];
 
     /**
      * @brief Gain of the first-order filter used for to update the magnetic field reference and candidate.
@@ -554,11 +554,11 @@ struct VQFCoefficients
     /**
      * @brief Numerator coefficients of the low-pass filter for the current magnetic norm and dip.
      */
-    double magNormDipLpB[3];
+    float magNormDipLpB[3];
     /**
      * @brief Denominator coefficients of the low-pass filter for the current magnetic norm and dip.
      */
-    double magNormDipLpA[2];
+    float magNormDipLpA[2];
 };
 
 /**
@@ -653,7 +653,7 @@ public:
      *
      * @param gyr gyroscope measurement in rad/s
      */
-    void updateGyr(const vqf_real_t gyr[3], double gyrTs);
+    void updateGyr(const vqf_real_t gyr[3], float gyrTs);
     /**
      * @brief Performs accelerometer update step.
      *
@@ -895,7 +895,7 @@ public:
      * @param outB output array for numerator coefficients
      * @param outA output array for denominator coefficients (without \f$a_0=1\f$)
      */
-    static void filterCoeffs(vqf_real_t tau, vqf_real_t Ts, double outB[3], double outA[2]);
+    static void filterCoeffs(vqf_real_t tau, vqf_real_t Ts, float outB[3], float outA[2]);
     /**
      * @brief Calculates the initial filter state for a given steady-state value.
      * @param x0 steady state value
@@ -903,7 +903,7 @@ public:
      * @param a denominator coefficients (without \f$a_0=1\f$)
      * @param out output array for filter state
      */
-    static void filterInitialState(vqf_real_t x0, const double b[], const double a[], double out[2]);
+    static void filterInitialState(vqf_real_t x0, const float b[], const float a[], float out[2]);
     /**
      * @brief Adjusts the filter state when changing coefficients.
      *
@@ -919,9 +919,9 @@ public:
      * @param a_new new denominator coefficients (without \f$a_0=1\f$)
      * @param state filter state (array of size N*2, will be modified)
      */
-    static void filterAdaptStateForCoeffChange(vqf_real_t last_y[], size_t N, const double b_old[3],
-                                               const double a_old[2], const double b_new[3],
-                                               const double a_new[2], double state[]);
+    static void filterAdaptStateForCoeffChange(vqf_real_t last_y[], size_t N, const float b_old[3],
+                                               const float a_old[2], const float b_new[3],
+                                               const float a_new[2], float state[]);
     /**
      * @brief Performs a filter step for a scalar value.
      * @param x input value
@@ -930,7 +930,7 @@ public:
      * @param state filter state array (will be modified)
      * @return filtered value
      */
-    static vqf_real_t filterStep(vqf_real_t x, const double b[3], const double a[2], double state[2]);
+    static vqf_real_t filterStep(vqf_real_t x, const float b[3], const float a[2], float state[2]);
     /**
      * @brief Performs filter step for vector-valued signal with averaging-based initialization.
      *
@@ -947,8 +947,8 @@ public:
      * @param state filter state (array of size N*2, will be modified)
      * @param out output array for filtered values (size N)
      */
-    static void filterVec(const vqf_real_t x[], size_t N, vqf_real_t tau, vqf_real_t Ts, const double b[3],
-                          const double a[2], double state[], vqf_real_t out[]);
+    static void filterVec(const vqf_real_t x[], size_t N, vqf_real_t tau, vqf_real_t Ts, const float b[3],
+                          const float a[2], float state[], vqf_real_t out[]);
 #ifndef VQF_NO_MOTION_BIAS_ESTIMATION
     /**
      * @brief Sets a 3x3 matrix to a scaled version of the identity matrix.
